+++ src/types/ingredient.types.ts
@@ -1,0 +1,11 @@
+export type IngredientData = Record<string, IngredientDataLabels>;
+
+export interface IngredientDataLabels {
+    en: string;
+    pl: string;
+}
+
+export interface IngredientManager {
+    labels: IngredientDataLabels[];
+    filterIngredients: (match: string) => IngredientDataLabels[];
+}